This PR is moving Swiftly to version 1.6.0. Hopefully everything will be
right this time.

Swiftly v1.6.0 will be under Pre-Release version until we see everything
going well

+++ src/server/menus/kinds/CenterMenu.h
@@ -3,7 +3,7 @@
 
 #include "../Menu.h"
 
-class CenterMenu: public Menu
+class CenterMenu : public Menu
 {
 private:
     std::string id;
@@ -25,7 +25,7 @@ public:
     void RegeneratePage(int playerid, int page, int selected);
 
     std::string GetCommandFromOption(int page, int selected);
-    size_t GetItemsOnPage(int page);
+    int GetItemsOnPage(int page);
     bool IsTemporary();
 
     Color GetColor();
